+++ back/src/main/java/fr/centralesupelec/thuv/dtos/SessionUpdateDto.java
@@ -2,10 +2,13 @@ package fr.centralesupelec.thuv.dtos;
 
 import lombok.Data;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
 
 @Data
 public class SessionUpdateDto {
     private Long id;
+    @Pattern(regexp = "^[^/]+$", message = "Title should not contain /")
+    private String title;
     @NotNull
     private Long startDateTime;
     @NotNull