+++ back/src/main/java/fr/centralesupelec/thuv/mappers/SessionMapper.java
@@ -28,6 +28,7 @@ public class SessionMapper {
                 .setId(
                         courseSession.getId()
                 )
+                .setTitle(courseSession.getTitle())
                 .setStartDateTime(
                         courseSession.getStartDateTime().atZone(zoneId).toInstant().toEpochMilli()
                 )
@@ -81,6 +82,7 @@ public class SessionMapper {
                 .setId(
                         courseSession.getId()
                 )
+                .setTitle(courseSession.getTitle())
                 .setStartDateTime(
                         courseSession.getStartDateTime().atZone(zoneId).toInstant().toEpochMilli()
                 )
@@ -107,6 +109,7 @@ public class SessionMapper {
             courseSession = new CourseSession();
         }
         return courseSession
+                .setTitle(courseSessionDto.getTitle())
                 .setStartDateTime(
                         Instant.ofEpochMilli(courseSessionDto.getStartDateTime()).atZone(zoneId).toLocalDateTime()
                 )