+++ back/src/main/java/fr/centralesupelec/thuv/model/CourseSession.java
@@ -20,6 +20,8 @@ public class CourseSession {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    private String title;
+
     @NotNull
     private LocalDateTime startDateTime;