+++ back/src/main/java/fr/centralesupelec/thuv/scale_up/mappers/CourseSessionWithResourceMapper.java
@@ -43,6 +43,7 @@ public class CourseSessionWithResourceMapper {
                 .setId(
                         courseSession.getId()
                 )
+                .setTitle(courseSession.getTitle())
                 .setStudentNumber(
                         courseSession.getStudentNumber()
                 )